+++ dotnet-desktop-guide/framework/wpf/advanced/snippets/threading-model/csharp/MultiWindow.xaml
@@ -1,0 +1,45 @@
+﻿<Window x:Class="SDKSamples.MultiWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        Title="Thread Hosted Window" Width="360" Height="180" SizeToContent="Height" ResizeMode="NoResize" Loaded="Window_Loaded">
+    <Grid>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="*" />
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="Auto" />
+        </Grid.RowDefinitions>
+
+        <TextBlock HorizontalAlignment="Right" Margin="30,0" Text="‼️" FontSize="50" FontWeight="ExtraBold"
+                   Foreground="Magenta" RenderTransformOrigin="0.5,0.5" Name="RotatedTextBlock">
+            <TextBlock.RenderTransform>
+                <RotateTransform Angle="0" />
+            </TextBlock.RenderTransform>
+            <TextBlock.Triggers>
+                <EventTrigger RoutedEvent="Loaded">
+                    <BeginStoryboard>
+                        <Storyboard>
+                            <DoubleAnimation Storyboard.TargetName="RotatedTextBlock"
+                                Storyboard.TargetProperty="(UIElement.RenderTransform).(RotateTransform.Angle)"
+                                From="0" To="360" Duration="0:0:5" RepeatBehavior="Forever" />
+                        </Storyboard>
+                    </BeginStoryboard>
+                </EventTrigger>
+            </TextBlock.Triggers>
+        </TextBlock>
+
+        <StackPanel Orientation="Horizontal" VerticalAlignment="Center" Margin="20" >
+            <Button Content="Pause" Click="StartStopButton_Click" Name="StartStopButton" Margin="5,0" Padding="10,0" />
+            <TextBlock Margin="5,0,0,0" Text="&lt;-- Pause for 5 seconds" />
+        </StackPanel>
+
+        <StackPanel Grid.Row="1" Margin="10">
+            <Button Content="Create 'Same Thread' Window" Click="SameThreadWindow_Click" />
+            <Button Content="Create 'New Thread' Window" Click="NewThreadWindow_Click" Margin="0,10,0,0" />
+        </StackPanel>
+
+        <StatusBar Grid.Row="2" VerticalAlignment="Bottom">
+            <StatusBarItem Content="Thread ID" Name="ThreadStatusItem" />
+        </StatusBar>
+
+    </Grid>
+</Window>
