+++ dotnet-desktop-guide/framework/wpf/advanced/snippets/threading-model/vb/Application.xaml
@@ -1,7 +1,7 @@
 ﻿<Application x:Class="Application"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             StartupUri="Weather.xaml">
+             StartupUri="MultiWindow.xaml">
     <Application.Resources>
 
         <!-- <clock_resources> -->
