--- dotnet-desktop-guide/samples/snippets/csharp/VS_Snippets_Wpf/ThreadingMultipleBrowsers/CSharp/Window1.xaml.cs
@@ -1,49 +1,0 @@
-﻿//<SnippetThreadingMultiBrowserCodeBehind>
-using System;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Threading;
-using System.Threading;
-
-namespace SDKSamples
-{
-    public partial class Window1 : Window
-    {
-
-        public Window1() : base()
-        {
-            InitializeComponent();
-        }
-
-        private void OnLoaded(object sender, RoutedEventArgs e)
-        {
-           placeHolder.Source = new Uri("http://www.msn.com");
-        }
-
-        private void Browse(object sender, RoutedEventArgs e)
-        {
-            placeHolder.Source = new Uri(newLocation.Text);
-        }
-
-        //<SnippetThreadingMultiBrowserNewWindow>
-        private void NewWindowHandler(object sender, RoutedEventArgs e)
-        {
-            Thread newWindowThread = new Thread(new ThreadStart(ThreadStartingPoint));
-            newWindowThread.SetApartmentState(ApartmentState.STA);
-            newWindowThread.IsBackground = true;
-            newWindowThread.Start();
-        }
-        //</SnippetThreadingMultiBrowserNewWindow>
-
-        //<SnippetThreadingMultiBrowserThreadStart>
-        private void ThreadStartingPoint()
-        {
-            Window1 tempWindow = new Window1();
-            tempWindow.Show();
-            System.Windows.Threading.Dispatcher.Run();
-        }
-        //</SnippetThreadingMultiBrowserThreadStart>
-    }
-}
-//</SnippetThreadingMultiBrowserCodeBehind>
