--- dotnet-desktop-guide/samples/snippets/csharp/VS_Snippets_Wpf/ThreadingMultipleBrowsers/CSharp/app.xaml
@@ -1,9 +1,0 @@
-﻿<Application x:Class="SDKSamples.app"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    StartupUri="Window1.xaml"
-    >
-    <Application.Resources>
-         
-    </Application.Resources>
-</Application>
