--- dotnet-desktop-guide/samples/snippets/csharp/VS_Snippets_Wpf/ThreadingPrimeNumbers/CSharp/Window1.xaml
@@ -1,17 +1,0 @@
-<!-- <SnippetThreadingPrimeNumberXAML> -->
-<Window x:Class="SDKSamples.Window1"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    Title="Prime Numbers" Width="260" Height="75"
-    >
-  <StackPanel Orientation="Horizontal" VerticalAlignment="Center" >
-    <Button Content="Start"  
-            Click="StartOrStop"
-            Name="startStopButton"
-            Margin="5,0,5,0"
-            />
-    <TextBlock Margin="10,5,0,0">Biggest Prime Found:</TextBlock>
-    <TextBlock Name="bigPrime" Margin="4,5,0,0">3</TextBlock>
-  </StackPanel>
-</Window>
-<!-- </SnippetThreadingPrimeNumberXAML> -->
