--- dotnet-desktop-guide/samples/snippets/csharp/VS_Snippets_Wpf/ThreadingWeatherForecast/CSharp/app.xaml.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Windows;
-using System.Data;
-using System.Xml;
-using System.Configuration;
-
-namespace SDKSamples
-{
-    /// <summary>
-    /// Interaction logic for app.xaml
-    /// </summary>
-
-    public partial class app : Application
-    {
-    }
-}
