--- dotnet-desktop-guide/samples/snippets/visualbasic/VS_Snippets_Wpf/ThreadingMultipleBrowsers/VisualBasic/Window1.xaml
@@ -1,23 +1,0 @@
-<Window x:Class="SDKSamples.Window1"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    Title="MultiBrowse"
-    Height="600" 
-    Width="800"
-    Loaded="OnLoaded"
-    >
-  <StackPanel Name="Stack" Orientation="Vertical">
-    <StackPanel Orientation="Horizontal">
-      <Button Content="New Window"
-              Click="NewWindowHandler" />
-      <TextBox Name="newLocation"
-               Width="500" />
-      <Button Content="GO!"
-              Click="Browse" />
-    </StackPanel>
-
-    <Frame Name="placeHolder"
-            Width="800"
-            Height="550"></Frame>
-  </StackPanel>
-</Window>
