--- dotnet-desktop-guide/samples/snippets/visualbasic/VS_Snippets_Wpf/ThreadingPrimeNumbers/visualbasic/application.xaml.vb
@@ -1,6 +1,0 @@
-﻿Class Application
-
-    ' Application-level events, such as Startup, Exit, and DispatcherUnhandledException
-    ' can be handled in this file.
-
-End Class
