--- dotnet-desktop-guide/samples/snippets/visualbasic/VS_Snippets_Wpf/ThreadingPrimeNumbers/visualbasic/mainwindow.xaml
@@ -1,15 +1,0 @@
-﻿<Window x:Class="SDKSamples.MainWindow"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    Title="Prime Numbers" Width="260" Height="75"
-    >
-    <StackPanel Orientation="Horizontal" VerticalAlignment="Center" >
-        <Button Content="Start"  
-            Click="StartOrStop"
-            Name="startStopButton"
-            Margin="5,0,5,0"
-            />
-        <TextBlock Margin="10,5,0,0">Biggest Prime Found:</TextBlock>
-        <TextBlock Name="bigPrime" Margin="4,5,0,0">3</TextBlock>
-    </StackPanel>
-</Window>
