--- dotnet-desktop-guide/samples/snippets/visualbasic/VS_Snippets_Wpf/ThreadingWeatherForecast/visualbasic/app.xaml.vb
@@ -1,15 +1,0 @@
-﻿Imports System.Windows
-Imports System.Data
-Imports System.Xml
-Imports System.Configuration
-
-Namespace SDKSamples
-	''' <summary>
-	''' Interaction logic for app.xaml
-	''' </summary>
-
-	Partial Public Class app
-		Inherits Application
-
-	End Class
-End Namespace
